Update bin2coff.c
The use of goto is not appropriate for the current software, and I noticed it in this source file. I only removed the goto statement; other than that, I did not see any errors in the code to the best of my ability. However, incorporating the if condition that was targeted by the goto statement into the code would make it more comprehensible.